Create SyllabusDownloadsServiceDelegate file
