+++ Stepic/Sources/Modules/CourseInfoSubmodules/CourseInfoTabSyllabus/Provider/SyllabusDownloadsService.swift
@@ -3,61 +3,6 @@ import PromiseKit
 
 // swiftlint:disable file_length
 
-// MARK: SyllabusDownloadsServiceDelegate -
-
-protocol SyllabusDownloadsServiceDelegate: AnyObject {
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveProgress progress: Float,
-        forVideoWithID videoID: Video.IdType
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveProgress progress: Float,
-        forImageURL url: URL
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveProgress progress: Float,
-        forUnitWithID unitID: Unit.IdType
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveProgress progress: Float,
-        forSectionWithID sectionID: Section.IdType
-    )
-
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveCompletion isCompleted: Bool,
-        forVideoWithID videoID: Video.IdType
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveCompletion isCompleted: Bool,
-        forImageURL url: URL
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveCompletion isCompleted: Bool,
-        forUnitWithID unitID: Unit.IdType
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didReceiveCompletion isCompleted: Bool,
-        forSectionWithID sectionID: Section.IdType
-    )
-
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didFailLoadVideoWithError error: Swift.Error
-    )
-    func syllabusDownloadsService(
-        _ service: SyllabusDownloadsServiceProtocol,
-        didFailLoadImageWithError error: Swift.Error
-    )
-}
-
 // MARK: - SyllabusDownloadsServiceProtocol -
 
 protocol SyllabusDownloadsServiceProtocol: AnyObject {
