+++ Stepic/Sources/Modules/CourseInfoSubmodules/CourseInfoTabSyllabus/Provider/SyllabusDownloadsServiceDelegate.swift
@@ -1,0 +1,127 @@
+import Foundation
+
+protocol SyllabusDownloadsServiceDelegate: AnyObject {
+    // MARK: On Progress Methods
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forVideoWithID videoID: Video.IdType
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forImageURL url: URL
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forUnitWithID unitID: Unit.IdType
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forSectionWithID sectionID: Section.IdType
+    )
+
+    // MARK: On Completion Methods
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forVideoWithID videoID: Video.IdType
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forImageURL url: URL
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forUnitWithID unitID: Unit.IdType
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forSectionWithID sectionID: Section.IdType
+    )
+
+    // MARK: On Error Methods
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didFailLoadVideoWithError error: Swift.Error
+    )
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didFailLoadImageWithError error: Swift.Error
+    )
+}
+
+extension SyllabusDownloadsServiceDelegate {
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forImageURL url: URL
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forUnitWithID unitID: Unit.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forVideoWithID videoID: Video.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveProgress progress: Float,
+        forSectionWithID sectionID: Section.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forVideoWithID videoID: Video.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forImageURL url: URL
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forUnitWithID unitID: Unit.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didReceiveCompletion isCompleted: Bool,
+        forSectionWithID sectionID: Section.IdType
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didFailLoadVideoWithError error: Error
+    ) {}
+
+    func syllabusDownloadsService(
+        _ service: SyllabusDownloadsServiceProtocol,
+        didFailLoadImageWithError error: Error
+    ) {}
+}
